+++ Settings.cs
@@ -6,7 +6,7 @@
 	public class Settings
 	{
 		public double FilterX = 1, FilterY = 1;
-		public uint Refresh = 0;
+		public ushort Refresh = 0;
 		public bool LinFit = true, Plot = true;
 		public string Xprop = "DataCorePlugin.GameData.AccelerationHeave",
 					  Yprop = "DataCorePlugin.GameData.GlobalAccelerationG";
